% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_na.R
\name{count_na}
\alias{count_na}
\title{Amount of missing values in a data set}
\usage{
count_na(data, percent = TRUE, col_sum = TRUE)
}
\arguments{
\item{data}{A data frame.}

\item{percent}{A logical value indicating whether to compute also the percentage (defaults to TRUE).}

\item{col_sum}{A logical value indicating whether the column sums should be returned.}
}
\value{
A tibble.
}
\description{
This function produces a table with the overall amount of missing values in a data set (absolute and relative values). #'
}
\examples{
d <- mtcars
d[4,3] <- NA # Create missing values to illustrate function

count_na(d, percent = FALSE)
count_na(d, col_sum = FALSE)
}
