% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_factor.r
\name{describe_factor}
\alias{describe_factor}
\title{Computes a frequency table for factor variables}
\usage{
describe_factor(x, name = NULL, labels = NULL)
}
\arguments{
\item{x}{A numeric or factor variable.}

\item{name}{A name for the first colum (if not specified, it remains the variable code).}

\item{labels}{If you are evaluating a single vector, you specify the variable name here.}
}
\description{
This function takes a factor variable (numeric variables are transformed into a factor variable) and computes basic frequency table.
}
\examples{
x <- rep(c("female", "male"), 100)

describe_factor(x, name = "gender")
}
