% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fcoplot.R
\name{fcoplot}
\alias{fcoplot}
\title{Scatter plot of x and y divided by z}
\usage{
fcoplot(x,y,z,varN)
}
\arguments{
\item{x}{:x-axis value}

\item{y}{:y-axis value}

\item{z}{:classification variable used to divided x and y}

\item{varN}{:variable name of z}
}
\description{
Generate 4 scatter plots of x and y divided by variable z, with a fitted line using a robust linear regression method.
}
\examples{
 fcoplot(mtcars[,1], mtcars[,3], mtcars[,4], "hp")
}
