% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdesstat.R
\name{fdesstat}
\alias{fdesstat}
\title{Descriptice statistics of a data frame}
\usage{
fdesstat(x,n)
}
\arguments{
\item{x}{:a data frame}

\item{n}{:number of decimal points desired}
}
\description{
Calculating the descriptive statistics of a data frame and exporting in a data frame.
The report data frame contains: number of observations, maximum value, minimum value, mean, median,
mode, variance, standard deviation, skewness and kurtosis.
}
\examples{
fdesstat(mtcars,3)
}
