% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdiff.R
\name{fdiff}
\alias{fdiff}
\title{Calculating the difference of a time series}
\usage{
fdiff(x,n)
}
\arguments{
\item{x}{: a numeric vector}

\item{n}{: number of lag periods}
}
\description{
Calculate the difference of a time series, with a specific lag period. The difference is used to show the
change in value over set period.
}
\examples{
fdiff(mtcars[,2],1)
}
