% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdrawdown.R
\name{fdrawdown}
\alias{fdrawdown}
\title{Largest draw down of returns}
\usage{
fdrawdown(x)
}
\arguments{
\item{x}{: a numeric vector of returns}
}
\description{
Calculate largest draw down of a series of returns. This function calculates the maximum
decrease in percentage over time, which can be used to test portfolio returns.
}
\examples{
fdrawdown(rnorm(100))
}
