% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flag.R
\name{flag}
\alias{flag}
\title{Create a lag variable}
\usage{
flag(x,n)
}
\arguments{
\item{x}{:a vector}

\item{n}{:number of lag periods}
}
\description{
Create a lag variable, with a choice of lag periods. The lag variable can be used
to test lag effects between variables.
}
\examples{
flag(mtcars[,2],3)
}
