% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flagcor.R
\name{flagcor}
\alias{flagcor}
\title{Lag/Lead Correlation}
\usage{
flagcor(x,y,lag,lead)
}
\arguments{
\item{x}{:the moving  vector}

\item{y}{:the fixed vector}

\item{lag}{:number of lag periods}

\item{lead}{:number of lead periods}
}
\description{
Calculating correlation of two vectors with lag and lead periods. The correlations are
used to determine the lag or lead effect between two variables. The corelattion function uses "na.or.complete"
method and calculate the pearson's correlation.
}
\examples{
flagcor(mtcars[,1],mtcars[,2],3,3)
}
