% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flead.R
\name{flead}
\alias{flead}
\title{Create a lead variable}
\usage{
flead(x,n)
}
\arguments{
\item{x}{:a vector}

\item{n}{:number of lead periods}
}
\description{
Create a lead variable, with a choice of lead periods. The lead variable can be used
to test lead effects between variables.
}
\examples{
flead(mtcars[,2],3)
}
