% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/flog.R
\name{flog}
\alias{flog}
\title{Sigmoid function}
\usage{
flog(x,top,a,b)
}
\arguments{
\item{x}{:a numeric vector}

\item{top}{:a numeric value as vertical scaler}

\item{a}{:a number to control top acceleration of the curve}

\item{b}{:a number to control bottom acceleration of the curve}
}
\description{
Generate sigmoid curve series, which is a specific case of
logistic function, with a control of top and bottom acceleration.
}
\examples{
 sigc <- round(flog(seq(-3, 3, 0.1), 1, -3, 3), 3)
ts.plot(sigc)
}
