% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fnd.R
\name{fnd}
\alias{fnd}
\title{Normal density function}
\usage{
fnd(x,mu,sig)
}
\arguments{
\item{x}{:x value}

\item{mu}{:mean value}

\item{sig}{:standard deviation}
}
\description{
Calculate normal density function value at x with a mean of mu and standard deviation of sig.
}
\examples{
 fnd(seq(-3, 3, 0.1), 0, 1)
}
