% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpctcng.R
\name{fpctcng}
\alias{fpctcng}
\title{Calculating rate of return of a vector}
\usage{
fpctcng(x,n)
}
\arguments{
\item{x}{:a numeric vector}

\item{n}{: number of lag periods}
}
\description{
Calculating the rate of return of a vector for further analysis, including calculating beta of companies, plotting to see the trend of the stock for technical analysis.
}
\examples{
fpctcng(mtcars[,1],1)
}
