% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fplot3.R
\name{fplot3}
\alias{fplot3}
\title{Scatter plot with text overlay}
\usage{
fplot3(x,y,txt)
}
\arguments{
\item{x}{: a numeric vector}

\item{y}{: a numeric vector}

\item{txt}{: a vector used as labels}
}
\description{
Generate a scatter plot with text overlay. This plot is to better show the effect of
the text variable in the domain of x and y variable.
}
\examples{
fplot3(mtcars[,1], mtcars[,3], row.names(mtcars))
}
