% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fplot3ss.R
\name{fplot3ss}
\alias{fplot3ss}
\title{Scatter smooth plot with text overlay}
\usage{
fplot3ss(x,y,txt,ce)
}
\arguments{
\item{x}{: a numeric vector}

\item{y}{: a numeric vector}

\item{txt}{: a vector used as labels}

\item{ce}{: text size, which default is set as 0.5}
}
\description{
Generate a scatter plot with text overlay, with a smooth curve fitted by loess.
}
\examples{
fplot3ss(mtcars[,1], mtcars[,3], row.names(mtcars))
}
