% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fsortcol.R
\name{fsortcol}
\alias{fsortcol}
\title{Sort a data frame by a column}
\usage{
fsortcol(x,n,dec)
}
\arguments{
\item{x}{:a data frame}

\item{n}{:number column to sort}

\item{dec}{:the order of sorting, default set to TRUE}
}
\description{
Sort a data frame by a column of choice. The column of choice is specified by the number of the column.
}
\examples{
fsortcol(mtcars,1,dec = TRUE)
}
