% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fspearcor.R
\name{fspearcor}
\alias{fspearcor}
\title{Spearman Rank Correlation}
\usage{
fspearcor(x,y)
}
\arguments{
\item{x}{:a numeric varibale}

\item{y}{:a numeric varibale}
}
\description{
Calculate Spearman Rank Correlation, which is the nonparametric version of the Pearson product-moment correlation.
}
\examples{
 fspearcor(mtcars[,1], mtcars[,3])
}
