% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funliov.R
\name{funliov}
\alias{funliov}
\title{Unite normal loss integral}
\usage{
funliov(x,disc,vol)
}
\arguments{
\item{x}{:a xtor}

\item{disc}{:discontinuity}

\item{vol}{:volatility}
}
\description{
Compute the value of the unit normal loss integral, with set discontinuity and volatility.
}
\examples{
 funliov(-3:10, 1, 3)
}
