% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist_var_plot.r
\name{hist_var_plot}
\alias{hist_var_plot}
\title{Create histograms for several Likert-type variables}
\usage{
hist_var_plot(data, bins = 5, fill = "#5e999b", color = "white",
  density = TRUE)
}
\arguments{
\item{data}{A data frame including all variables that should be plotted.}

\item{bins}{Number of values that should be plotted on the x-axis (should equal the number of answer options)}

\item{fill}{Color of the bins.}

\item{color}{Color of the bin margins.}

\item{density}{A logical value indicating whether a normally distributed density curve should be plotted on top of the histogram. The density curve takes the mean and sd of the respective variable as arguments. In this case, the histograms are computed using \code{aes(y = stat(density)}.}
}
\value{
An object of class \code{tibble} that can be passed to \code{apa_table()}. It contains all sample sizes that were provided as individual numeric values.
}
\description{
This functions computes a facetted plot of variable distributions. Using the density function, the first layer is a histogram and the second a normal distributed density curve based on the mean and standard deviation of the respective variable.
}
\examples{
d <- data.frame(x = rnorm(100, 3, 1),
y = rnorm(100, 3, 1))

hist_var_plot(d, density = F)
hist_var_plot(d)
}
