% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invert.r
\name{invert}
\alias{invert}
\title{Invert 5- or 7-point scales}
\usage{
invert(x, length = 5)
}
\arguments{
\item{x}{The variables that should be inverted}

\item{length}{Is it a 5-point of 7-point scale?}
}
\value{
A vector
}
\description{
This function serves as a wrapper for \code{recode} from the "car" package. It allows to easily invert 5- or 7-point scales.
}
