% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_pattern_plot.r
\name{missing_pattern_plot}
\alias{missing_pattern_plot}
\title{Plotting missingness patterns}
\usage{
missing_pattern_plot(data, colors = c("#2F6FAF", "lightblue"),
  frequency = TRUE, ratio = c(2.5, 1), nrow = 1)
}
\arguments{
\item{data}{A data frame.}

\item{colors}{Colors to be used in the plot.}

\item{frequency}{A logical value indicating whether the frequency of the missingness pattens should be plotted to}

\item{ratio}{Vector specifying the size of both plots in comparison to one another.}

\item{nrow}{Underneath or next to each other?}
}
\value{
A gtable.
}
\description{
This function plots the missingness patterns and if wanted also their their frequencies.
}
\examples{
d <- mtcars
d[4,3] <- NA # Create missing to illustrate function

missing_pattern_plot(d)
}
