% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi_join.r
\name{multi_join}
\alias{multi_join}
\title{Customized version of the function inner_join() (dplyr) for more than two data frames}
\usage{
multi_join(...)
}
\arguments{
\item{...}{Up to five data sets that should be joined.}
}
\value{
A data frame.
}
\description{
This function joins up to five data sets using the function \code{inner_join} from the package "dplyr". Each data set is individually joined with the previous one. This way, several data sets of which only pairs have the same id variable can be joined.
}
\examples{
# Creating several data frames.
id <- c(1:100)
var1 <- runif(100, 1, 5)
var2 <- runif(100, 1, 5)
var3 <- runif(100, 1, 5)
data1 <- data.frame(id, var1)
data2 <- data.frame(id, var2)
data3 <- data.frame(id, var3)

# Join all of them
multi_join(data1, data2, data3)
}
