% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pl.histgg.R
\name{pl.histgg}
\alias{pl.histgg}
\title{Plot histograms for a data frame with ggplot2}
\usage{
pl.histgg(x, l = 1)
}
\arguments{
\item{x}{:a data frame}

\item{l}{: the beginning label number in the title (default set to 1)}
}
\description{
Plotting histograms for a data frame with 4 per image, with titles and label numbers automatically generated.
}
\note{
This function uses Freedman-Diaconis rule for histogram bin size calculation. In some occasions feed discrete variables to this function
may yield terrible results.
}
\examples{
pl.histgg(as.data.frame(EuStockMarkets),1)
}
\references{
Freedman, David; Diaconis, Persi (December 1981). "On the histogram as a density estimator: L2 theory" (PDF). Probability Theory and Related Fields. Heidelberg: Springer Berlin. 57 (4): 453–476. doi:10.1007/BF01025868. ISSN 0178-8051. Retrieved 2009-01-06.
}
