% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pl.multiplot.R
\name{pl.multiplot}
\alias{pl.multiplot}
\title{Multiple  plot function for ggplot2 objects}
\usage{
pl.multiplot(..., plotlist = NULL, cols = 1, layout = NULL)
}
\arguments{
\item{...}{:ggplot2 objects}

\item{plotlist}{:a list of ggplot2 objects}

\item{cols}{:number of columns in the layout}

\item{layout}{:a matrix that specifies the layout. It has higher priority than 'cols'}
}
\description{
Render multiple ggplt2 plots on one page
}
\references{
https://stackoverflow.com/questions/24387376/r-error-could-not-find-function-multiplot-using-cookbook-example
}
