% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reg.linreg.R
\name{reg.linreg}
\alias{reg.linreg}
\title{Linear regression processor}
\usage{
reg.linreg(dataframe,dependent)
}
\arguments{
\item{dataframe}{:a data frame, which includes the dependent variable}

\item{dependent}{:dependent variable}
}
\description{
This function will take a data frame and the dependent variable, and fit all possible combinations of models.
The result will be a data frame of models and test statistics for all the models possible.
}
\examples{
reg.linreg(mtcars,"mpg")
}
