% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_sizes.r
\name{sample_sizes}
\alias{sample_sizes}
\title{Creates table with sample sizes and attrition rates}
\usage{
sample_sizes(..., sample = c("Overall clicks", "Eligible cases",
  "Complete cases"))
}
\arguments{
\item{...}{Several numeric values representing sample sizes (e.g., obtained by using \code{save_n()}}

\item{sample}{Names for the rows in the final table.}
}
\value{
An object of class \code{tibble} that can be passed to \code{apa_table()}. It contains all sample sizes that were provided as individual numeric values.
}
\description{
This functions binds individual numeric values (representing sample sizes) in a printable table. It is a follow-up function for \code{save_n} which extracts the number of cases in a data frame.
}
\examples{
# Simple example
sample_sizes(100, 80, 10)

# Realistic example
d <- mtcars
d1 <- d[1:28,]
d2 <- d1[1:20,]

n0 <- save_n(d)
n1 <- save_n(d1)
n2 <- save_n(d2)

sample_sizes(n0, n1, n2)
}
