% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/socdem_table.r
\name{socdem_table}
\alias{socdem_table}
\title{Socio-demographics table}
\usage{
socdem_table(age, gender, edu, gender_labels = NULL, edu_labels = NULL,
  frequency = TRUE, percent = TRUE)
}
\arguments{
\item{age}{A numeric variable representing the age of the participants.}

\item{gender}{A numeric or factor variable representing the gender of the participants.}

\item{edu}{A numeric or factor variable representing the educational level of the participants.}

\item{gender_labels}{If gender is of class "numeric", you can provide a vector with factor labels.}

\item{edu_labels}{If edu is of class "numeric", you can provide a vector with factor labels.}

\item{frequency}{Logical value indicating whether frequencies should be returned.}

\item{percent}{Logical value indicating whether percentages should be returned.}
}
\value{
The table summarizing the socio-demographic characteristics of a data frame.
}
\description{
This function takes socio-demographic variables and transforms them into a printable table. A numeric variable representing age (in years) will be summarized as mean and standard deviation. Gender and education will be returned as frequency tables.
}
\examples{
# Creating variables
age <- c(20, 21, 25, 35, 19, 20)
gender <- c(rep("female", 3), rep("male", 3))
edu <- c(rep("highschool", 2), rep("middleschool", 4))

# Summarizing
(table_1 <- socdem_table(age, gender, edu))

# Rmarkdown table
papaja::apa_table(
  table_1,
  caption = "Socio-Demographics",
  stub_indents = list("Gender" = c(2,3),
                      "Education" = c(4,5))
)
}
