% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/count_na.R
\name{count_na}
\alias{count_na}
\title{Count overall amount of missing values}
\usage{
count_na(data, percent = TRUE)
}
\arguments{
\item{data}{A data frame.}

\item{percent}{A logical value indicating whether to compute also the percentage (defaults to TRUE).}
}
\value{
A tibble displaying the number and (if \code{percent = TRUE}) percent of missings in the data frame.
}
\description{
This function counts the amount of missing values in a data set.
}
\examples{
d <- mtcars
d[4,3] <- NA # Create missing to illustrate function

count_na(d, percent = FALSE)
count_na(d)
}
