% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_factor.r
\name{describe_factor}
\alias{describe_factor}
\title{Computes a frequency table for factor variables}
\usage{
describe_factor(x, name = NULL, labels = NULL, col_sum = TRUE, ...)
}
\arguments{
\item{x}{A numeric or factor variable.}

\item{name}{A name for the first colum (if not specified, it remains the variable code).}

\item{labels}{If you are evaluating a numeric vector, the function will transform it into a factor variable. You can specify custom labels for the factor levels here.}

\item{col_sum}{Do you want column sums?}

\item{...}{Further arguments that can be passed to \code{table}. This has been primarily included to allow to count NAs. In this case, simply add the following argument: \code{useNA = "always"}}
}
\value{
A tibble.
}
\description{
This function takes a factor variable (numeric variables are transformed into a factor variable) and computes a basic frequency table.
}
\examples{
# Standard example
d <- mtcars
describe_factor(d$cyl, name = "No. of cylinders")

# Example with missing values
d$cyl[3] <- NA # Creating a missing value
describe_factor(d$cyl, name = "No. of cylinders", useNA = "always")
}
