% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe_factor.r
\name{describe_factor}
\alias{describe_factor}
\title{A frequency table for factor variables}
\usage{
describe_factor(x, name = NULL, labels = NULL, cum_percent = FALSE,
  col_sum = TRUE, view = FALSE, ...)
}
\arguments{
\item{x}{A numeric or factor variable.}

\item{name}{A name for the first colum (if not specified, it remains the variable code).}

\item{labels}{If you are evaluating a numeric vector, the function will transform it into a factor variable. You can specify custom labels for the factor levels here.}

\item{cum_percent}{Logical value indicating whether cumulative percent should also be returned.}

\item{col_sum}{Logical value indicating whether column sums should be returned.}

\item{view}{Logical value indicating whether the full data frame should be printed (defaults to FALSE).}

\item{...}{Further arguments that can be passed to \code{\link[base]{table}}. This has been primarily included to allow to count NAs. In this case, simply add the following argument: \code{useNA = "always"}}
}
\value{
The function \code{describe_factor()} returns a contigency table with relative and absolute values and an object of class \code{tbl_df} (a data frame).
}
\description{
This function takes a factor variable (numeric variables are transformed into a factor variable) and computes a basic frequency table. Column sums are added automatically, but can be removed if desired. By default, NAs will be removed before computation. Yet, arguments that are usally passed to the \code{\link[base]{table}} function can be added to also count NAs (useNA = "always").
}
\examples{
# Standard example
d <- mtcars
describe_factor(d$cyl, name = "No. of cylinders")
describe_factor(d$mpg, view = TRUE)

# Example with missing values
d$cyl[3] <- NA # Creating a missing value
describe_factor(d$cyl, 
                name = "No. of cylinders", 
                cum_percent = TRUE,
                useNA = "always")
}
