% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hist_var_plot.r
\name{hist_var_plot}
\alias{hist_var_plot}
\title{Create histograms for several Likert-type variables}
\usage{
hist_var_plot(data, bins = 5, fill = "#5e999b", color = "white",
  density = TRUE, time = NULL, key = NULL)
}
\arguments{
\item{data}{An object of one of the following classes: "numeric", "double", "integer", "data.frame", "tbl", or "tbl_df".}

\item{bins}{Number of values that should be plotted on the x-axis (should equal the number of answer options or a reasonable scale).}

\item{fill}{Color of the bins (defaults to green).}

\item{color}{Color of the bin borders (defaults to white).}

\item{density}{A logical value indicating whether a normally distributed density curve should be plotted on top of the histogram. The density curve takes the mean and sd of the respective variables as arguments. In this case, the histograms are computed using \code{aes(y = stat(density)}.}

\item{time}{If the data is panel data, this argument can be used to provide the time variable. Yet, in this case, the data has to be transformed into the long format. This argument only works in combination with the following argument "key" (see examples).}

\item{key}{If the data is panel data, this argument can be used to provide the key variable (variable that indicates the variable names in the long format). Again, in this case, the data has to be transformed into the long format. This argument only works in combination with the following argument "time" (see examples).}
}
\value{
A \code{\link[ggplot2]{ggplot}} object that can be further customized using standard ggplot2 elements.
}
\description{
This functions computes a facetted plot of variable distributions. Using the density function, the first layer is a histogram and the second a normal distributed density curve based on the mean and standard deviation of the respective variable.
}
\examples{
## Example 1: Standard usage

d <- data.frame(x = rnorm(100, 3, 1),
                y = rnorm(100, 3, 1))

# Several variables
hist_var_plot(d, density = F)
hist_var_plot(d)

# One variables
hist_var_plot(d$x)

## Example 2: Panel data

d <- data.frame(T1_var1 = rnorm(100, 3, 1),
                T2_var1 = rnorm(100, 3.5, 1),
                T1_var2 = rnorm(100, 2, 1),
                T2_var2 = rnorm(100, 1.5, 1))

# Transforming data into long format
d \%>\% 
  gather(key, value) \%>\%
  separate(key, c("time", "key"), sep = 2) \%>\%
  hist_var_plot(., time = time, key = code)
}
