% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invert.r
\name{invert}
\alias{invert}
\title{Invert scales with any number of answer options}
\usage{
invert(x, length = 5)
}
\arguments{
\item{x}{The variable that should be inverted (needs to be numeric!).}

\item{length}{Number of answer options (defaults to )}
}
\value{
A vector representing the recoded variable.
}
\description{
This function allows to easily invert items with any number of answer options.
}
