% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/missing_pattern_plot.r
\name{missing_pattern_plot}
\alias{missing_pattern_plot}
\title{Missingness pattern plot}
\usage{
missing_pattern_plot(data, colors = c("#2F6FAF", "lightblue"),
  titles = c("A", "B"), var_labels = FALSE, frequency = TRUE,
  percent = FALSE, ratio = c(2.5, 1), nrow = 1)
}
\arguments{
\item{data}{The data frame that should be evaluated.}

\item{colors}{Colors to be used in the plot.}

\item{titles}{Both plots can be named individually. By default, they are simply labelled "A" and "B".}

\item{var_labels}{A logical value indicating whether variable names should be plotted (if many variables are in the data frame, setting this to TRUE can be messy).}

\item{frequency}{A logical value indicating whether the frequency of the missingness pattens should be plotted to.}

\item{percent}{A logical value indicating whether the share of cases that have a particular missingness pattern should be plotted into the frequency table (still beta, doesn't look very nice..., simply helps to understand the data)}

\item{ratio}{Vector specifying the size of both plots in comparison to one another.}

\item{nrow}{Should the plots be printed underneath or next to each other? Defaults to "next to each other".}
}
\value{
A \code{gtable} which consists by default of two plots: (A) An overview of all existing missingness patterns in the data frame, (B) a frequency table representing how often the particular missingness patterns can be found in the data frame.
}
\description{
This function plots the missingness patterns and (if desired) also their their frequencies in the data frame.
}
\examples{
d <- mtcars
d[4,3] <- NA # Create missing to illustrate function

missing_pattern_plot(d)

# Add variable names and plot only patterns
missing_pattern_plot(d, var_labels = T, frequency = F)
}
