% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save_n.r
\name{save_n}
\alias{save_n}
\title{Save numer of cases as numeric value}
\usage{
save_n(data)
}
\arguments{
\item{data}{A data frame.}
}
\value{
A numeric value.
}
\description{
Just a convenient function to save the number of cases of a data frame. This function primarily serves to provide values for the wrapper \code{sample_sizes()} in this package.
}
\examples{
d <- mtcars
save_n(d)
}
